Replace deprecated method innovation of `hasAnnotationsOf()`
This commit replaces the innovation of the `hasAnnotationsOf()` that would be removed in `V1.0.0` with the appropriate alternatives:

- `hasAnnotationOf()`
- `hasAllAnnotationsOf()`

So anybody using the library examples will not use code examples that are going to be replaces in the relatively near future.

---

Moreover, it changes the library calls as well to use the appropriate function internally instead of the deprecated version.