Add file information to getKtFile IllegalArgumentException
We've encountered issues in our CI due to some bogus file in the custom scope we were creating, in a flakey test. 

But finding the offending file is harder than it needs to be, as the exception won't say which file it is.

I'm adding the file path to the error, to know earlier which file caused the issues. It's also consistent with the exception being thrown in L49 in KotlinFileParser.kt.