Optimize loading of KoFiles
The current implementation has two issues, which can cause issues
on larger projects:

- Currently getKoFileDeclarations() will parse _all_ Kotlin files in
  the project, including build files. They then later filtered out by
  KoScopeCreatorCore.getFiles, but this is very inefficient.
- KoFileDeclarationProvider will try and launch `getKoFileDeclarations`
  from object init (which is basically when the class is loaded).
  I assume that this was done to preload the classes, but I think that
  is very aggressive. These should be lazily loaded when they're first needed,
  which will usually be the first test.

This PR changes this so that `getKoFileDeclarations` can be provided with
a filter block, which is provided by `KoScopeCreatorCore.getFiles`.

Also optimized all of the Regex creation. We don't want to create these on
every file in the tree-walk (Regex instances are quite heavy).